Refactor FormBuilder
