+++ src/components/Navbar/Navbar.tsx
@@ -5,14 +5,14 @@ import Btn from '../Btn/Btn';
 import MoreIcon from '../../images/icons/ellipsis-horizontal-outline.svg';
 import useOutsideClick from '../../hooks/useOutsideClick';
 import './Navbar.css';
+import PublishModal from '../PublishModal/PublishModal';
+import JSONView from '../JSONView/JSONView';
+import PredefinedValueSetModal from '../PredefinedValueSetModal/PredefinedValueSetModal';
+import ImportValueSet from '../ImportValueSet/ImportValueSet';
 
 type Props = {
     newQuestionnaire: () => void;
-    showAdmin: () => void;
     showFormFiller: () => void;
-    showJSONView: () => void;
-    showImportValueSet: () => void;
-    showContained: () => void;
     uploadQuestionnaire: (event: React.ChangeEvent<HTMLInputElement>) => void;
 };
 
@@ -22,17 +22,13 @@ enum MenuItem {
     more = 'more',
 }
 
-const Navbar = ({
-    newQuestionnaire,
-    showAdmin,
-    showFormFiller,
-    showJSONView,
-    showImportValueSet,
-    showContained,
-    uploadQuestionnaire,
-}: Props): JSX.Element => {
+const Navbar = ({ newQuestionnaire, showFormFiller, uploadQuestionnaire }: Props): JSX.Element => {
     const { state } = useContext(TreeContext);
     const [selectedMenuItem, setSelectedMenuItem] = useState(MenuItem.none);
+    const [showContained, setShowContained] = useState(false);
+    const [showImportValueSet, setShowImportValueSet] = useState(false);
+    const [showJSONView, setShowJSONView] = useState(false);
+    const [showPublish, setShowPublish] = useState(false);
     const navBarRef = useRef<HTMLHeadingElement>(null);
 
     useOutsideClick(navBarRef, () => {
@@ -87,54 +83,63 @@ const Navbar = ({
     };
 
     return (
-        <header ref={navBarRef}>
-            <div>
-                <Btn title="Skjema" onClick={() => handleMenuItemClick(MenuItem.file)} />
-                {selectedMenuItem === MenuItem.file && (
-                    <div className="menu file">
-                        <Btn title="Nytt skjema" onClick={() => callbackAndHide(newQuestionnaire)} />
-                        <label className="regular-btn upload-btn">
-                            <input
-                                type="file"
-                                style={{ display: 'none' }}
-                                onChange={(event) => {
-                                    uploadQuestionnaire(event);
-                                    hideMenu();
-                                }}
-                                accept="application/JSON"
-                            />
-                            Last opp skjema
-                        </label>
+        <>
+            <header ref={navBarRef}>
+                <div>
+                    <Btn title="Skjema" onClick={() => handleMenuItemClick(MenuItem.file)} />
+                    {selectedMenuItem === MenuItem.file && (
+                        <div className="menu file">
+                            <Btn title="Nytt skjema" onClick={() => callbackAndHide(newQuestionnaire)} />
+                            <label className="regular-btn upload-btn">
+                                <input
+                                    type="file"
+                                    style={{ display: 'none' }}
+                                    onChange={(event) => {
+                                        uploadQuestionnaire(event);
+                                        hideMenu();
+                                    }}
+                                    accept="application/JSON"
+                                />
+                                Last opp skjema
+                            </label>
+                        </div>
+                    )}
+                </div>
+
+                <div className="left"></div>
+
+                <div className="pull-right">
+                    <Btn title="Forhåndsvisning" onClick={showFormFiller} />
+                    <Btn title="Lagre" onClick={() => exportToJsonAndDownload()} />
+                    <div
+                        className="more-menu"
+                        tabIndex={0}
+                        role="button"
+                        aria-label="menu list"
+                        aria-pressed="false"
+                        onClick={() => handleMenuItemClick(MenuItem.more)}
+                        onKeyPress={(e) => e.code === 'Enter' && handleMenuItemClick(MenuItem.more)}
+                    >
+                        <img className="more-menu-icon" src={MoreIcon} alt="more icon" height={25} />
+                    </div>
+                </div>
+                {selectedMenuItem === MenuItem.more && (
+                    <div className="menu">
+                        <Btn title="JSON" onClick={() => callbackAndHide(() => setShowJSONView(!showJSONView))} />
+                        <Btn title="Publiser" onClick={() => callbackAndHide(() => setShowPublish(!showPublish))} />
+                        <Btn
+                            title="Importer valg"
+                            onClick={() => callbackAndHide(() => setShowImportValueSet(!showImportValueSet))}
+                        />
+                        <Btn title="Valg" onClick={() => callbackAndHide(() => setShowContained(!showContained))} />
                     </div>
                 )}
-            </div>
-
-            <div className="left"></div>
-
-            <div className="pull-right">
-                <Btn title="Forhåndsvisning" onClick={showFormFiller} />
-                <Btn title="Lagre" onClick={() => exportToJsonAndDownload()} />
-                <div
-                    className="more-menu"
-                    tabIndex={0}
-                    role="button"
-                    aria-label="menu list"
-                    aria-pressed="false"
-                    onClick={() => handleMenuItemClick(MenuItem.more)}
-                    onKeyPress={(e) => e.code === 'Enter' && handleMenuItemClick(MenuItem.more)}
-                >
-                    <img className="more-menu-icon" src={MoreIcon} alt="more icon" height={25} />
-                </div>
-            </div>
-            {selectedMenuItem === MenuItem.more && (
-                <div className="menu">
-                    <Btn title="JSON" onClick={() => callbackAndHide(showJSONView)} />
-                    <Btn title="Publiser" onClick={() => callbackAndHide(showAdmin)} />
-                    <Btn title="Importer valg" onClick={() => callbackAndHide(showImportValueSet)} />
-                    <Btn title="Valg" onClick={() => callbackAndHide(showContained)} />
-                </div>
-            )}
-        </header>
+            </header>
+            {showContained && <PredefinedValueSetModal close={() => setShowContained(!showContained)} />}
+            {showImportValueSet && <ImportValueSet close={() => setShowImportValueSet(!showImportValueSet)} />}
+            {showJSONView && <JSONView showJSONView={() => setShowJSONView(!showJSONView)} />}
+            {showPublish && <PublishModal close={() => setShowPublish(!showPublish)} />}
+        </>
     );
 };
