--- src/views/MainMenu.css
@@ -1,52 +1,0 @@
-.main-header {
-    background: #fff;
-    padding: 0 2.5rem;
-    border-bottom: 1px solid #eaeaec;
-}
-
-.mainheader-top {
-    color: #2c3845;
-    height: 4rem;
-    display: flex;
-    align-items: center;
-    max-width: 75rem;
-    margin: 0 auto;
-}
-
-.logo {
-    padding-top: 8px;
-    width: 100px;
-}
-
-.app-name {
-    margin-left: auto;
-}
-
-.align-everything {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-}
-
-.main-menu {
-    background: #f0f0f0;
-    min-height: calc(100vh - 65px);
-}
-
-.main-menu > div {
-    background-color: white;
-    height: 40vh;
-    width: 750px;
-    box-shadow: rgba(0, 0, 0, 0.1) 40px 40px 40px;
-}
-
-.main-menu label.index-buttons {
-    margin-left: 15px;
-}
-
-.main-menu img.upload-icon {
-    margin: 4px;
-    width: 25px;
-    height: 25px;
-    filter: invert();
-}
